fixer uper

+++ includes/class-wc-ajax.php
@@ -65,7 +65,8 @@ class WC_AJAX {
 			'term_ordering'                                    => false,
 			'product_ordering'                                 => false,
 			'refund_line_items'                                => false,
-			'delete_refund'                                    => false
+			'delete_refund'                                    => false,
+			'rated'                                            => false
 		);
 
 		foreach ( $ajax_events as $ajax_event => $nopriv ) {
@@ -95,7 +96,7 @@ class WC_AJAX {
 					'div.widget_shopping_cart_content' => '<div class="widget_shopping_cart_content">' . $mini_cart . '</div>'
 				)
 			),
-			'cart_hash' => apply_filters( 'woocommerce_add_to_cart_hash', WC()->cart->get_cart() ? md5( json_encode( WC()->cart->get_cart() ) ) : '', WC()->cart->get_cart() )
+			'cart_hash' => apply_filters( 'woocommerce_add_to_cart_hash', WC()->cart->get_cart_for_session() ? md5( json_encode( WC()->cart->get_cart_for_session() ) ) : '', WC()->cart->get_cart_for_session() )
 		);
 
 		wp_send_json( $data );
@@ -414,6 +415,10 @@ class WC_AJAX {
 
 		check_ajax_referer( 'add-attribute', 'security' );
 
+		if ( ! current_user_can( 'edit_products' ) ) {
+			die(-1);
+		}
+
 		global $wc_product_attributes;
 
 		$thepostid     = 0;
@@ -424,7 +429,7 @@ class WC_AJAX {
 		$attribute     = array(
 			'name'         => $taxonomy,
 			'value'        => '',
-			'is_visible'   => 1,
+			'is_visible'   => apply_filters( 'woocommerce_attribute_default_visibility', 1 ),
 			'is_variation' => 0,
 			'is_taxonomy'  => $taxonomy ? 1 : 0
 		);
@@ -450,8 +455,12 @@ class WC_AJAX {
 
 		check_ajax_referer( 'add-attribute', 'security' );
 
+		if ( ! current_user_can( 'manage_product_terms' ) ) {
+			die(-1);
+		}
+
 		$taxonomy = esc_attr( $_POST['taxonomy'] );
-		$term     = stripslashes( $_POST['term'] );
+		$term     = wc_clean( $_POST['term'] );
 
 		if ( taxonomy_exists( $taxonomy ) ) {
 
@@ -462,10 +471,11 @@ class WC_AJAX {
 					'error' => $result->get_error_message()
 				) );
 			} else {
+				$term = get_term_by( 'id', $result['term_id'], $taxonomy );
 				wp_send_json( array(
-					'term_id' => $result['term_id'],
-					'name'    => $term,
-					'slug'    => sanitize_title( $term ),
+					'term_id' => $term->term_id,
+					'name'    => $term->name,
+					'slug'    => $term->slug
 				) );
 			}
 		}
@@ -478,6 +488,10 @@ class WC_AJAX {
 	 */
 	public static function remove_variations() {
 		check_ajax_referer( 'delete-variations', 'security' );
+
+		if ( ! current_user_can( 'edit_products' ) ) {
+			die(-1);
+		}
 
 		$variation_ids = (array) $_POST['variation_ids'];
 
@@ -498,6 +512,10 @@ class WC_AJAX {
 	public static function save_attributes() {
 
 		check_ajax_referer( 'save-attributes', 'security' );
+
+		if ( ! current_user_can( 'edit_products' ) ) {
+			die(-1);
+		}
 
 		// Get post data
 		parse_str( $_POST['data'], $data );
@@ -611,6 +629,10 @@ class WC_AJAX {
 
 		check_ajax_referer( 'add-variation', 'security' );
 
+		if ( ! current_user_can( 'edit_products' ) ) {
+			die(-1);
+		}
+
 		$post_id = intval( $_POST['post_id'] );
 		$loop = intval( $_POST['loop'] );
 
@@ -714,6 +736,10 @@ class WC_AJAX {
 		}
 
 		check_ajax_referer( 'link-variations', 'security' );
+
+		if ( ! current_user_can( 'edit_products' ) ) {
+			die(-1);
+		}
 
 		@set_time_limit(0);
 
@@ -862,7 +888,7 @@ class WC_AJAX {
 			}
 		}
 
-		delete_transient( 'wc_product_children_ids_' . $post_id );
+		delete_transient( 'wc_product_children_ids_' . $post_id . WC_Cache_Helper::get_transient_version( 'product' ) );
 
 		echo $added;
 
@@ -875,6 +901,10 @@ class WC_AJAX {
 	public static function revoke_access_to_download() {
 
 		check_ajax_referer( 'revoke-access', 'security' );
+
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
 
 		global $wpdb;
 
@@ -895,6 +925,10 @@ class WC_AJAX {
 	public static function grant_access_to_download() {
 
 		check_ajax_referer( 'grant-access', 'security' );
+
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
 
 		global $wpdb;
 
@@ -950,6 +984,10 @@ class WC_AJAX {
 
 		check_ajax_referer( 'get-customer-details', 'security' );
 
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
+
 		$user_id      = (int) trim(stripslashes($_POST['user_id']));
 		$type_to_load = esc_attr(trim(stripslashes($_POST['type_to_load'])));
 
@@ -970,38 +1008,17 @@ class WC_AJAX {
 		$customer_data = apply_filters( 'woocommerce_found_customer_details', $customer_data, $user_id, $type_to_load );
 
 		wp_send_json( $customer_data );
-
 	}
 
 	/**
 	 * Add order item via ajax
 	 */
 	public static function add_order_item() {
-
-		function webseo_prices( $backend_prod_id ){
-
-			$pluginpath = "/home/devsitesco/public_html/wp-content/plugins/codecanyon-7104096-woo-table-based-pricing/";
-			include_once $pluginpath . 'woocommerce-price-table.php';
-			$height = $_POST['width']; //posted from html-order-items.php
-			$width = $_POST['height'];
-			
-			$prices = get_prices( $width, $height, $backend_prod_id );
-			$json_string = json_encode( $prices ); //json encode prices
-			$obj = json_decode( $json_string, true );
-			return $obj['product_price'];
-			/*foreach ( $obj['addon_prices'] as $taxonomy => $addons ) {
-				foreach ( $addons as $addon => $addonPrices ) {
-					foreach ( $addonPrices as $addonPrice ) {
-						if ( ( float )$addonPrice > 0 ) {
-							return $addonPrice;
-						}
-					}
-				}
-			}*/
-			
-		}
-
 		check_ajax_referer( 'order-item', 'security' );
+
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
 
 		$item_to_add = sanitize_text_field( $_POST['item_to_add'] );
 		$order_id    = absint( $_POST['order_id'] );
@@ -1031,9 +1048,9 @@ class WC_AJAX {
 		$item['name']              = $_product->get_title();
 		$item['tax_class']         = $_product->get_tax_class();
 		$item['qty']               = 1;
-		$item['line_subtotal']     = webseo_prices( $_product->id );
+		$item['line_subtotal']     = wc_format_decimal( $_product->get_price_excluding_tax() );
 		$item['line_subtotal_tax'] = '';
-		$item['line_total']        = webseo_prices( $_product->id );
+		$item['line_total']        = wc_format_decimal( $_product->get_price_excluding_tax() );
 		$item['line_tax']          = '';
 
 		// Add line item
@@ -1081,6 +1098,10 @@ class WC_AJAX {
 
 		check_ajax_referer( 'order-item', 'security' );
 
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
+
 		$order_id      = absint( $_POST['order_id'] );
 		$order         = wc_get_order( $order_id );
 		$order_taxes   = $order->get_taxes();
@@ -1108,6 +1129,10 @@ class WC_AJAX {
 	public static function add_order_shipping() {
 
 		check_ajax_referer( 'order-item', 'security' );
+
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
 
 		$order_id         = absint( $_POST['order_id'] );
 		$order            = wc_get_order( $order_id );
@@ -1137,6 +1162,10 @@ class WC_AJAX {
 
 		check_ajax_referer( 'order-item', 'security' );
 
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
+
 		$order_id = absint( $_POST['order_id'] );
 		$rate_id  = absint( $_POST['rate_id'] );
 		$order    = wc_get_order( $order_id );
@@ -1156,6 +1185,10 @@ class WC_AJAX {
 	 */
 	public static function remove_order_item() {
 		check_ajax_referer( 'order-item', 'security' );
+
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
 
 		$order_item_ids = $_POST['order_item_ids'];
 
@@ -1179,6 +1212,10 @@ class WC_AJAX {
 
 		check_ajax_referer( 'order-item', 'security' );
 
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
+
 		$order_id = absint( $_POST['order_id'] );
 		$rate_id  = absint( $_POST['rate_id'] );
 
@@ -1197,6 +1234,10 @@ class WC_AJAX {
 	 */
 	public static function reduce_order_item_stock() {
 		check_ajax_referer( 'order-item', 'security' );
+
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
 
 		$order_id       = absint( $_POST['order_id'] );
 		$order_item_ids = isset( $_POST['order_item_ids'] ) ? $_POST['order_item_ids'] : array();
@@ -1243,6 +1284,10 @@ class WC_AJAX {
 	 */
 	public static function increase_order_item_stock() {
 		check_ajax_referer( 'order-item', 'security' );
+
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
 
 		$order_id       = absint( $_POST['order_id'] );
 		$order_item_ids = isset( $_POST['order_item_ids'] ) ? $_POST['order_item_ids'] : array();
@@ -1291,6 +1336,10 @@ class WC_AJAX {
 	public static function add_order_item_meta() {
 		check_ajax_referer( 'order-item', 'security' );
 
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
+
 		$meta_id = wc_add_order_item_meta( absint( $_POST['order_item_id'] ), __( 'Name', 'woocommerce' ), __( 'Value', 'woocommerce' ) );
 
 		if ( $meta_id ) {
@@ -1308,6 +1357,10 @@ class WC_AJAX {
 
 		check_ajax_referer( 'order-item', 'security' );
 
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
+
 		$meta_id = absint( $_POST['meta_id'] );
 
 		$wpdb->query( $wpdb->prepare( "DELETE FROM {$wpdb->prefix}woocommerce_order_itemmeta WHERE meta_id = %d", $meta_id ) );
@@ -1322,6 +1375,10 @@ class WC_AJAX {
 		global $wpdb;
 
 		check_ajax_referer( 'calc-totals', 'security' );
+
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
 
 		$tax            = new WC_Tax();
 		$order_id       = absint( $_POST['order_id'] );
@@ -1464,6 +1521,10 @@ class WC_AJAX {
 	public static function save_order_items() {
 		check_ajax_referer( 'order-item', 'security' );
 
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
+
 		if ( isset( $_POST['order_id'] ) && isset( $_POST['items'] ) ) {
 			$order_id = absint( $_POST['order_id'] );
 
@@ -1489,6 +1550,10 @@ class WC_AJAX {
 	public static function load_order_items() {
 		check_ajax_referer( 'order-item', 'security' );
 
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
+
 		// Return HTML items
 		$order_id = absint( $_POST['order_id'] );
 		$order    = wc_get_order( $order_id );
@@ -1505,7 +1570,11 @@ class WC_AJAX {
 
 		check_ajax_referer( 'add-order-note', 'security' );
 
-		$post_id   = (int) $_POST['post_id'];
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
+
+		$post_id   = absint( $_POST['post_id'] );
 		$note      = wp_kses_post( trim( stripslashes( $_POST['note'] ) ) );
 		$note_type = $_POST['note_type'];
 
@@ -1536,6 +1605,10 @@ class WC_AJAX {
 
 		check_ajax_referer( 'delete-order-note', 'security' );
 
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
+
 		$note_id = (int) $_POST['note_id'];
 
 		if ( $note_id > 0 ) {
@@ -1552,10 +1625,14 @@ class WC_AJAX {
 	 * @param string $x (default: '')
 	 * @param string $post_types (default: array('product'))
 	 */
-	public static function json_search_products( $x = '', $post_types = array('product') ) {
+	public static function json_search_products( $x = '', $post_types = array( 'product' ) ) {
 		ob_start();
 
 		check_ajax_referer( 'search-products', 'security' );
+
+		if ( ! current_user_can( 'edit_products' ) ) {
+			die(-1);
+		}
 
 		$term = (string) wc_clean( stripslashes( $_GET['term'] ) );
 
@@ -1563,17 +1640,25 @@ class WC_AJAX {
 			die();
 		}
 
+		$args = array(
+			'post_type'      => $post_types,
+			'post_status'    => 'publish',
+			'posts_per_page' => -1,
+			's'              => $term,
+			'fields'         => 'ids'
+		);
+
 		if ( is_numeric( $term ) ) {
 
-			$args = array(
+			$args2 = array(
 				'post_type'      => $post_types,
 				'post_status'    => 'publish',
 				'posts_per_page' => -1,
-				'post__in'       => array(0, $term),
+				'post__in'       => array( 0, $term ),
 				'fields'         => 'ids'
 			);
 
-			$args2 = array(
+			$args3 = array(
 				'post_type'      => $post_types,
 				'post_status'    => 'publish',
 				'posts_per_page' => -1,
@@ -1581,7 +1666,7 @@ class WC_AJAX {
 				'fields'         => 'ids'
 			);
 
-			$args3 = array(
+			$args4 = array(
 				'post_type'      => $post_types,
 				'post_status'    => 'publish',
 				'posts_per_page' => -1,
@@ -1595,17 +1680,9 @@ class WC_AJAX {
 				'fields'         => 'ids'
 			);
 
-			$posts = array_unique( array_merge( get_posts( $args ), get_posts( $args2 ), get_posts( $args3 ) ) );
+			$posts = array_unique( array_merge( get_posts( $args ), get_posts( $args2 ), get_posts( $args3 ), get_posts( $args4 ) ) );
 
 		} else {
-
-			$args = array(
-				'post_type'      => $post_types,
-				'post_status'    => 'publish',
-				'posts_per_page' => -1,
-				's'              => $term,
-				'fields'         => 'ids'
-			);
 
 			$args2 = array(
 				'post_type'      => $post_types,
@@ -1631,7 +1708,7 @@ class WC_AJAX {
 			foreach ( $posts as $post ) {
 				$product = wc_get_product( $post );
 
-				$found_products[ $post ] = $product->get_formatted_name();
+				$found_products[ $post ] = rawurldecode( $product->get_formatted_name() );
 			}
 		}
 
@@ -1649,7 +1726,7 @@ class WC_AJAX {
 	 * @see WC_AJAX::json_search_products()
 	 */
 	public static function json_search_products_and_variations() {
-		self::json_search_products( '', array('product', 'product_variation') );
+		self::json_search_products( '', array( 'product', 'product_variation' ) );
 	}
 
 	/**
@@ -1659,6 +1736,10 @@ class WC_AJAX {
 		ob_start();
 
 		check_ajax_referer( 'search-customers', 'security' );
+
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
 
 		$term = wc_clean( stripslashes( $_GET['term'] ) );
 
@@ -1700,6 +1781,12 @@ class WC_AJAX {
 	 */
 	public static function json_search_downloadable_products_and_variations() {
 		ob_start();
+
+		check_ajax_referer( 'search-products', 'security' );
+
+		if ( ! current_user_can( 'edit_products' ) ) {
+			die(-1);
+		}
 
 		$term = (string) wc_clean( stripslashes( $_GET['term'] ) );
 
@@ -1755,10 +1842,16 @@ class WC_AJAX {
 	 * Ajax request handling for categories ordering
 	 */
 	public static function term_ordering() {
+
+		// check permissions again and make sure we have what we need
+		if ( ! current_user_can( 'edit_products' ) || empty( $_POST['id'] ) ) {
+			die(-1);
+		}
+
 		$id       = (int) $_POST['id'];
 		$next_id  = isset( $_POST['nextid'] ) && (int) $_POST['nextid'] ? (int) $_POST['nextid'] : null;
 		$taxonomy = isset( $_POST['thetaxonomy'] ) ? esc_attr( $_POST['thetaxonomy'] ) : null;
-		$term     = get_term_by('id', $id, $taxonomy);
+		$term     = get_term_by( 'id', $id, $taxonomy );
 
 		if ( ! $id || ! $term || ! $taxonomy ) {
 			die(0);
@@ -1872,6 +1965,10 @@ class WC_AJAX {
 
 		check_ajax_referer( 'order-item', 'security' );
 
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
+
 		$order_id               = absint( $_POST['order_id'] );
 		$refund_amount          = wc_format_decimal( sanitize_text_field( $_POST['refund_amount'] ) );
 		$refund_reason          = sanitize_text_field( $_POST['refund_reason'] );
@@ -1887,7 +1984,7 @@ class WC_AJAX {
 			// Validate that the refund can occur
 			$order       = wc_get_order( $order_id );
 			$order_items = $order->get_items();
-			$max_refund  = $order->get_total() - $order->get_total_refunded();
+			$max_refund  = wc_format_decimal( $order->get_total() - $order->get_total_refunded() );
 
 			if ( ! $refund_amount || $max_refund < $refund_amount ) {
 				throw new exception( __( 'Invalid refund amount', 'woocommerce' ) );
@@ -1984,13 +2081,31 @@ class WC_AJAX {
 	public static function delete_refund() {
 		check_ajax_referer( 'order-item', 'security' );
 
+		if ( ! current_user_can( 'edit_shop_orders' ) ) {
+			die(-1);
+		}
+
 		$refund_id = absint( $_POST['refund_id'] );
 
 		if ( $refund_id && 'shop_order_refund' === get_post_type( $refund_id ) ) {
 			wc_delete_shop_order_transients( wp_get_post_parent_id( $refund_id ) );
 			wp_delete_post( $refund_id );
+
+			do_action( 'woocommerce_refund_deleted', $refund_id );
 		}
 
+		die();
+	}
+
+	/**
+	 * Triggered when clicking the rating footer.
+	 */
+	public static function rated() {
+		if ( ! current_user_can( 'manage_woocommerce' ) ) {
+			die(-1);
+		}
+
+		update_option( 'woocommerce_admin_footer_text_rated', 1 );
 		die();
 	}
 }
